% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_confidence_interval.R
\name{calculate_CI_from_error_distribution}
\alias{calculate_CI_from_error_distribution}
\title{Get confidence interval for ancestral or hidden states}
\usage{
calculate_CI_from_error_distribution(
  error,
  alpha = 0.05,
  tol = .Machine$double.eps^0.25
)
}
\arguments{
\item{error}{a data frame listing the mean, variance and weight of normal distributions}

\item{alpha}{the desired confidence level;default to 0.05}

\item{tol}{the precision of calculated CI that is passed to \code{uniroot} with the same default}
}
\description{
Calculate the confidence interval from the calculated posterior probability
}
\details{
\code{calculate_CI_from_error_distribution} calls \code{uniroot} function to find the CI at the desired confidence level.

For a mixture of normal distribution, the quantile at cumulative probability p Q_p will always be within the interval

\link{min(Q_p_i),max(Q_p_i)} where Q_p_i stands for the quantiles at cumulative probability p for each of the normal component.
}
