% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_Gaussian_decomposition.R
\name{calculate_error_mean_and_var}
\alias{calculate_error_mean_and_var}
\title{Calculate error mean value}
\usage{
calculate_error_mean_and_var(error)
}
\description{
Calculate error mean value
}
