% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_Gaussian_decomposition.R
\name{calculate_error_mean_and_var}
\alias{calculate_error_mean_and_var}
\title{Calculate the mean and variance of an error distribution}
\usage{
calculate_error_mean_and_var(error)
}
\arguments{
\item{error}{an error distributions (a data frame)}
}
\value{
A named vector of two containing the mean and variance of the error distribution
}
\description{
The function \link{calculate_error_mean_and_var} calculates the mean and variance of an error distribution
}
