% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_diagnosis.R
\name{calculateExclusion}
\alias{calculateExclusion}
\title{Calculate tolerance-based inclusion/exclusion}
\usage{
calculateExclusion(
  pred,
  error,
  epsilon = 0,
  tolerance,
  alpha = 0.05,
  laplace = FALSE,
  discrete = FALSE
)
}
\description{
Calculate tolerance-based inclusion/exclusion
}
