% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_diagnosis.R
\name{calculateExclusionErrorRate}
\alias{calculateExclusionErrorRate}
\title{Calcualte the error rate}
\usage{
calculateExclusionErrorRate(
  obs,
  pred,
  error,
  epsilon = 0,
  tolerance,
  alpha = 0.05,
  laplace = FALSE,
  discrete = FALSE
)
}
\description{
Calcualte the error rate
}
