% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_diagnosis.R
\name{calculateHeteroscedasticity}
\alias{calculateHeteroscedasticity}
\title{Calculate median Chi-square}
\usage{
calculateHeteroscedasticity(x, y, bin = 100, n = 1000)
}
\description{
Calculate median Chi-square
}
