% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_diagnosis.R
\name{checkExclusion}
\alias{checkExclusion}
\title{Check if exclusion/inclusion are correct}
\usage{
checkExclusion(obs, pred, tolerance, discrete = FALSE)
}
\description{
Check if exclusion/inclusion are correct
}
