% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_reconstruction_utils.R
\name{compoundNormalMoments}
\alias{compoundNormalMoments}
\title{Calculate four central moments of a mixture of normal distributions}
\usage{
compoundNormalMoments(probs, means, vars)
}
\arguments{
\item{probs}{the vector of probability weight of normal distributions}

\item{means}{the vector of means of mixed normal distributions}

\item{vars}{the vector of standard deviations of mixed normal distributions}
}
\value{
A numeric vector of length 4 (mean, variance, skewness and kurtosis)
}
\description{
Calculate the mean, variance, skewness and excess kurtosis of the mixture of normal distributions
}
