% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_global_ancestral_state.R
\name{continuousAceByPIC}
\alias{continuousAceByPIC}
\title{Predict global ancestral state under the BM model}
\usage{
continuousAceByPIC(x, phy, rate = 1)
}
\arguments{
\item{x}{named vector of tip trait values; names should match the tip labels with no missing values}

\item{phy}{phylo-class object from \code{ape} package}

\item{rate}{the rate of BM}
}
\value{
a data frame listing the means and variances of the ancestral states
}
\description{
Predict global ancestral state under the BM model given a phylogeny, known tip states, and the rate of BM
}
