% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_diagnosis_utils.R
\name{discretizeResult}
\alias{discretizeResult}
\title{Discretize result}
\usage{
discretizeResult(res, error = NULL, epsilon = 0, laplace = FALSE, numCores = 1)
}
\description{
Extending RasperGade to integer traits
}
