% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_diagnosis_utils.R
\name{F_beta_score}
\alias{F_beta_score}
\alias{precision_recall}
\title{Calculating recall, precision and F score}
\usage{
F_beta_score(obs, pred, beta = 1)

precision_recall(obs, pred, as.na = FALSE)
}
\description{
Functions to calculate recall, precision, and the F score with variable weight between recall and precision
}
