% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_normal_approximation.R
\name{findBestNormalApproximation1}
\alias{findBestNormalApproximation1}
\alias{findBestNormalApproximation2}
\title{Find approximation for a mixture of normal distributions using limited number of normal distributions}
\usage{
findBestNormalApproximation1(probs, means, vars)

findBestNormalApproximation2(probs, means, vars, numCores = 1)
}
\arguments{
\item{probs}{the vector of probability weight of normal distributions}

\item{means}{the vector of means of mixed normal distributions}

\item{vars}{the vector of standard deviations of mixed normal distributions}
}
\value{
A dataframe with 3 columns, with columns representing the means, variances and probability weights of the normal distributions to approximate the full mixture
}
\description{
Find a solution of using 1 or 2 normal distributions to approximate a mixture of normal distributions
}
