% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_utils.R
\name{findSisters}
\alias{findSisters}
\title{Find the sister nodes/tips of a node/tip}
\usage{
findSisters(phy, x)
}
\arguments{
\item{phy}{a phylo-class object from the \code{ape} package}

\item{x}{the index of the focal node}
}
\value{
The indices of the sisters. If the root is supplied, `NA`` will be returned.

If a non-root node has no sister, an empty numeric vector will be returned.
}
\description{
An utility function to find the sister nodes/tips of a node/tip
}
\seealso{
[getRoot][getAncestor][getNextDescendants][getConnected]
}
