% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_fitting.R
\name{findTerminalPIC}
\alias{findTerminalPIC}
\title{Find only tip-pair contrasts in a tree}
\usage{
findTerminalPIC(phy, x, remove.zero = FALSE)
}
\description{
Find only tip-pair contrasts in a tree for a reconstruction-free analysis
}
