% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_fitting.R
\name{fit.model.byNM.2step}
\alias{fit.model.byNM.2step}
\alias{fit.model.byNM}
\title{Wrapper function for optimization}
\usage{
fit.model.byNM.2step(
  fit.func = "mle2",
  start.params,
  params1,
  params2,
  AIC.func = "AIC",
  params.func = function(x) {     x@coef },
  ini.func = "as.list",
  AIC.cutoff = 2,
  ...
)

fit.model.byNM(
  fit.func = "mle2",
  start.params,
  params,
  AIC.func = "AIC",
  params.func = function(x) {     x@coef },
  ini.func = "as.list",
  AIC.cutoff = 2,
  ...
)
}
\description{
Fit pulsed evolution with reliable results
}
