% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_fitting.R
\name{fitPE}
\alias{fitPE}
\alias{fitPE.phy}
\title{Fit pulsed evolution model}
\usage{
fitPE(
  x,
  l,
  start.value = NULL,
  min.value = rep(0, 4),
  numCores = 1,
  fixed = list(),
  laplace = TRUE,
  eval.only = FALSE,
  reltol = sqrt(.Machine$double.eps),
  ...
)

fitPE.phy(
  phy,
  x,
  start.value = NULL,
  min.value = rep(0, 4),
  numCores = 1,
  fixed = list(),
  ignore.zero = TRUE,
  laplace = TRUE,
  eval.only = FALSE,
  bootstrap = NULL,
  reltol = sqrt(.Machine$double.eps),
  ...
)
}
\arguments{
\item{x}{trait change in \code{fitPE} and tip trait values in \code{fitPE.phy}}

\item{l}{branch length of time/divergence}

\item{phy}{a phylo-class object from the \code{ape} package}
}
\description{
Fit pulsed evolution model in a maximum-likelihood framework
}
