% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_reconstruction.R
\name{fullMarginalReconstructionWithPE}
\alias{fullMarginalReconstructionWithPE}
\title{Reconstruct all marginal ancestral state in a phylogeny}
\usage{
fullMarginalReconstructionWithPE(
  phy,
  x,
  params,
  laplace = FALSE,
  approximate = 1,
  margin = 1e-06,
  numCores = 1,
  asymptotic = 5
)
}
\arguments{
\item{phy}{phylo-class object from \code{ape} package}

\item{x}{named vector of tip trait values; names should match the tip labels; NA should not be included}

\item{params}{a vector of pulsed evolution parameters}

\item{laplace}{logical, when true, time-independent variation follows Laplace distribution}

\item{approximate}{number of normal distributions to approximate the uncertainty distribution for internal nodes}

\item{margin}{total probability mass of the number of jump not taken into account per compound Poisson process}

\item{numCores}{number of cores to use}

\item{asymptotic}{threshold of expected number of jumps beyond which the trait change is assumed to be normal}
}
\value{
returns a list that contains all the information to calculate the global ancestral states and hidden states
}
\description{
Reconstruct all marginal ancestral states, including those require rerooting, in a phylogeny
}
\details{
Parameter vector should be in the format of lambda,size,...,sigma,epsilon
Currently \code{laplace} should always set to be \code{FALSE}.
}
