% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_utils.R
\name{getLatestAncestor}
\alias{getLatestAncestor}
\alias{getAncestor}
\title{Find the ancestor of a node or tip}
\usage{
getLatestAncestor(phy, x)

getAncestor(phy, x, level = 1, trace = FALSE)
}
\arguments{
\item{phy}{a phylo-class object from the \code{ape} package}

\item{x}{the index of the focal node/tip}

\item{level}{the depth between the ancestor and the focal node/tip}

\item{trace}{logical, if FALSE (the default), only the earliest ancestor is returned, otherwise, all ancestors along the path are returned}
}
\value{
A vector of ancestor indices is returned. If root is supplied, \code{NA} will be returned
}
\description{
Utility functions to find the ancestors of a node or tip
}
\seealso{
[getRoot][getNextDescendants][findSisters][getConnected]
}
