% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_utils.R
\name{findEdges}
\alias{findEdges}
\alias{getConnected}
\title{Find the edges/nodes connected to a node}
\usage{
findEdges(phy, node, ancestor = FALSE)

getConnected(phy, x)
}
\arguments{
\item{phy}{a phylo-class object from the \code{ape} package}

\item{node}{the index of the focal node}

\item{ancestor}{logical, if FALSE (the default), the edges originating from the node are returned, otherwise, the edge leading to the node is returned.}

\item{x}{the index of the focal node}
}
\value{
\code{findEdges} returns the indices of the edges. Note that these indices are different from those of nodes/tips

\code{getConnected} returns the indices of the connected nodes.
}
\description{
An utility function to find the sister nodes/tips of a node/tip
}
\seealso{
[getRoot][getAncestor][getNextDescendants][findSisters]
}
