% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_fitting_utils.R
\name{get.params}
\alias{get.params}
\alias{get.AIC}
\title{Get a component of the fitted model}
\usage{
get.params(x)

get.AIC(x)
}
\description{
Get a component of the fitted model from the list of results
}
