% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_utils.R
\name{getNextDescendants}
\alias{getNextDescendants}
\title{Find the immediate descendants of a node}
\usage{
getNextDescendants(phy, x)
}
\arguments{
\item{phy}{a phylo-class object from the \code{ape} package}

\item{x}{the index of the focal node}
}
\value{
The indices of the immediate descendants. If a tip is supplied, `NA`` will be returned.
}
\description{
An utility function to find the immediate descendants of a node
}
\seealso{
[getRoot][getAncestor][findSisters][getConnected]
}
