% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_utils.R
\name{getRoot}
\alias{getRoot}
\title{Find the root of a phylogeny}
\usage{
getRoot(phy)
}
\arguments{
\item{phy}{a phylo-class object from the \code{ape} package}
}
\value{
The index of the root
}
\description{
An utility function to find the root of a tree
}
\seealso{
[getAncestor][getNextDescendants][findSisters][getConnected]
}
