% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_fitting_utils.R
\name{initialize.parameters.SP}
\alias{initialize.parameters.SP}
\title{Prepare model parameters for optimization}
\usage{
initialize.parameters.SP(x, ...)
}
\description{
Prepare the parameter vector so it can be supplied to the optimizer
}
