% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_node_tracking.R
\name{locate_query_sequences}
\alias{locate_query_sequences}
\title{Locate query sequences}
\usage{
locate_query_sequences(query.phy, query.tips)
}
\description{
Find the position of insertion in a phylogeny with query tips inserted
}
