% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_diagnosis_utils.R
\name{logitSigmoid}
\alias{logitSigmoid}
\alias{reverse.logitSigmoid}
\title{Shift the center of deviations to a probability}
\usage{
logitSigmoid(x, p)

reverse.logitSigmoid(x, p)
}
\description{
Shift the center of deviations to a probability for a balanced measurement
}
