% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_cross_validation.R
\name{LOO_CV_with_BM}
\alias{LOO_CV_with_BM}
\title{Leave-one-out cross-validation under BM model}
\usage{
LOO_CV_with_BM(phy, trait)
}
\arguments{
\item{phy}{phylo-class object from \code{ape} package}

\item{trait}{named vector of tip trait values; names should match the tip labels; missing values are hidden states}

\item{numCores}{number of threads to run in parallel}
}
\value{
\verb{$summary} is a data frame listing the means and variances of the hidden states

\verb{$error} is a list of error distributions where each element is a data frame
}
\description{
Predict hidden states under the BM model in a leave-one-out cross-validation
}
