% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_node_tracking.R
\name{make_traceable_edge_label}
\alias{make_traceable_edge_label}
\title{Calculating traceable node labels}
\usage{
make_traceable_edge_label(phy)
}
\description{
Calculate a traceable node label that can be maintained among trees
}
