% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_reconstruction.R
\name{marginalReconstruction}
\alias{marginalReconstruction}
\alias{marginalReconstructionWithPE}
\title{Reconstruct marginal ancestral state from a node's immediate descendants}
\usage{
marginalReconstruction(x, l)

marginalReconstructionWithPE(
  x,
  l,
  lambdas,
  sizes,
  sigma,
  epsilons,
  margin = 1e-06,
  numCores = 1,
  asymptotic = 5
)
}
\arguments{
\item{x}{vector of means of mixed normal distributions}

\item{l}{vector of standard deviations of mixed normal distributions}

\item{lambdas}{vector of jump frequencies, has the same length as the number of compound Poisson processes}

\item{sizes}{vector of variances of trait change per jump, has the same length as the number of compound Poisson processes}

\item{sigma}{rate of BM, has a length of 1}

\item{epsilons}{vector of uncertainty not explained by the branches, has the same length as the number of immediate descendants}

\item{margin}{total probability mass of the number of jump not taken into account per compound Poisson process}

\item{numCores}{number of cores to use}

\item{asymptotic}{threshold of expected number of jumps beyond which the trait change is assumed to be normal}
}
\value{
\code{marginalReconstruction} returns a vector of the mean variance and the scaling factor

\code{marginalReconstructionWithPE} returns a list of the mean, variance and a data frame of the normal mixture
}
\description{
Reconstruct marginal ancestral state using immediate descendants under BM or pulsed evolution model
}
