% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_Gaussian_decomposition.R
\name{mix_errors_by_weight}
\alias{mix_errors_by_weight}
\title{Mix error distributions}
\usage{
mix_errors_by_weight(errors, weights)
}
\description{
Mix error distributions by weights
}
