% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_node_tracking.R
\name{partition_test_set_with_minimum_distance}
\alias{partition_test_set_with_minimum_distance}
\title{Partition test set from reference data}
\usage{
partition_test_set_with_minimum_distance(
  phy,
  trait,
  test.set,
  distance = 0,
  location = TRUE
)
}
\description{
Prepare the reference data for cross-validation at different degree of uncertainty
}
