% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEpois.R
\name{dPEpoisnorm}
\alias{dPEpoisnorm}
\alias{pPEpoisnorm}
\alias{rPEpoisnorm}
\title{Probability functions of trait change under pulsed evolution with 1 compound Poisson process}
\usage{
dPEpoisnorm(x, t, lambda, size, sigma, epsilon = 0, margin = 1e-06)

pPEpoisnorm(
  q,
  t,
  lambda,
  size,
  sigma,
  epsilon = 0,
  margin = 1e-06,
  lower.tail = TRUE
)

rPEpoisnorm(n, t, lambda, size, sigma, epsilon = 0)
}
\arguments{
\item{x}{the trait change between two nodes}

\item{t}{the branch length between two nodes}

\item{lambda}{the rate at which jumps happen}

\item{size}{the variance of trait change introduced per jump}

\item{sigma}{the rate of BM (variance of gradual trait change per unit branch length)}

\item{epsilon}{the variance of time-independent variation}

\item{margin}{the total probability mass of the number of jumps not covered, defaults to 1e-6}

\item{q}{the trait change between two nodes}

\item{lower.tail}{logical, if TRUE (default), probability of trait change smaller or equal to q is returned. If false, the probability of being greater is returned.}

\item{n}{the number of random values to return}
}
\value{
\code{dPEpoisnorm} returns the probability density

\code{pPEpoisnorm} returns the cumulative distribution

\code{rPEpoisnorm} returns a vector of random numbers following the specified distribution
}
\description{
Density function, cumulative distribution function and random generation for the trait change under Pulsed evolution with 1 compound Poisson process
}
\examples{
# generate some random trait changes under pulsed evolution
dtrait = rPEpoisnorm(n=1e3,t=1,lambda=1,size=1,sigma=1,epsilon=0.1)

# test the empirical distribution against the theoretical one
ks.test(dtrait,pPEpoisnorm,t=1,lambda=1,size=1,sigma=1,epsilon=0.1)
}
