% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_alias.R,
%   R/RasperGade_hidden_state_prediction.R
\name{predTraitByPIC}
\alias{predTraitByPIC}
\alias{predict_hidden_state_by_pic}
\title{Predict hidden state under the BM model}
\usage{
predTraitByPIC(phy, trait)

predict_hidden_state_by_pic(phy, trait)
}
\arguments{
\item{phy}{phylo-class object from \code{ape} package}

\item{trait}{named vector of tip trait values; names should match the tip labels; missing values (NA or not present) are hidden states}
}
\value{
a list of two elements:

\verb{$summary} is a data frame listing the means and variances of the hidden states

\verb{$error} is a list of error distributions where each element is a data frame
}
\description{
Predict hidden state under the BM model given a phylogeny, some known states, and the rate of BM
}
