% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_hidden_state_prediction.R
\name{predictHiddenStateWithPE}
\alias{predictHiddenStateWithPE}
\title{Predict hidden state under the PE model}
\usage{
predictHiddenStateWithPE(
  FMR,
  query.keys,
  laplace = FALSE,
  numApprox = 1,
  margin = 1e-06,
  asymptotic = 5,
  numCores = 1
)
}
\arguments{
\item{FMR}{the returned data structure from function \code{fullMarginalReconstructionWithPE}}

\item{query.keys}{the tracking key of query tips}

\item{laplace}{logical, if true, Laplace distribution is used for time-independent variation}

\item{numApprox}{the number of normal distributions to approximate the Laplace distribution}

\item{margin}{the total probability mass that the number of jumps omitted in a compound Poisson process}

\item{asymptotic}{the threshold of expected number of jumps on a branch beyond which normal distribution is assumed}

\item{numCores}{the number of cores to run in parallel}
}
\value{
a list of two elements:

\verb{$summary} is a data frame listing the means and variances of the hidden states

\verb{$error} is a list of error distributions#' @details to be added
}
\description{
Predict hidden state under the PE model given a phylogeny, some known states, and the PE model
}
