% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_hidden_state_prediction.R
\name{predTraitByPIC}
\alias{predTraitByPIC}
\title{Predict hidden state under the BM model}
\usage{
predTraitByPIC(phy, trait, rate = 1)
}
\arguments{
\item{phy}{phylo-class object from \code{ape} package}

\item{trait}{named vector of tip trait values; names should match the tip labels; hidden states to be predicted have NA values}

\item{rate}{the rate of BM}
}
\description{
Reconstruct marginal ancestral state using immediate descendants under BM or pulsed evolution model
}
\details{
this function is modified from the function in \code{picante} package
}
