% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_fitting_utils.R
\name{total.process.variance}
\alias{total.process.variance}
\title{Total evolution rate}
\usage{
total.process.variance(params)
}
\description{
Calculate the long-term evolution rate combining BM and PE
}
