% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_reconstruction_utils.R
\name{productNormalDensity}
\alias{productNormalDensity}
\title{Calculate the product of normal probability density functions}
\usage{
productNormalDensity(mean, var)
}
\arguments{
\item{mean}{the vector of means of mixed normal distributions}

\item{var}{the vector of standard deviations of mixed normal distributions}
}
\value{
A numeric vector of length 3 (mean, variance and scale)
}
\description{
Calculate the mean, variance and the scaling factor of the product of normal pdfs (a Gaussian function)
}
\details{
The function is implemented based on \href{http://www.lucamartino.altervista.org/2003-003.pdf}{Bromiley 2014}
}
