% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_diagnosis_utils.R
\name{pseudo.Z.score}
\alias{pseudo.Z.score}
\title{Calculate pseudo-Z-score}
\usage{
pseudo.Z.score(trait, pred, error, epsilon = 0, laplace = FALSE)
}
\description{
Calculate pseudo-Z-score that follows the standard normal distribution
}
