% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_fitting.R
\name{reconstructAncestralStates}
\alias{reconstructAncestralStates}
\title{Reconstruct ancestral states under BM and time-independent variation}
\usage{
reconstructAncestralStates(phy, x, rate, epsilon)
}
\description{
Reconstruct ancestral states under BM and time-independent variation for optimization of model parameters
}
