% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_utils.R
\name{reroot_tree_at_tip}
\alias{reroot_tree_at_tip}
\title{Root the phylogeny at a tip}
\usage{
reroot_tree_at_tip(phy, tip)
}
\arguments{
\item{phy}{a phylo-class object from the \code{ape} package}

\item{tip}{the focal tip, either a character string matching the tip label or a number matching its index}
}
\value{
A phylo-class object containing the re-rooted tree
}
\description{
An utility function to root the tree at a tip for hidden state prediction through ancestral reconstruction
}
