% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_simulation.R
\name{simulatePET}
\alias{simulatePET}
\alias{simulatePET.phy}
\title{Simulate trait under pulsed evolution}
\usage{
simulatePET(l, align.length = NA, lambda, size, sigma, epsilon, nj = NULL)

simulatePET.phy(phy, lambda, size, sigma, epsilon, start = 0)
}
\arguments{
\item{l}{branch length of time/divergence}

\item{lambda}{jump frequency}

\item{size}{variance of trait change per jump}

\item{sigma}{rate of BM}

\item{epsilon}{variance of time-independent variation}

\item{phy}{a phylo-class object from the \code{ape} package}
}
\description{
Simulate trait under pulsed evolution
}
