% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_fitting_utils.R
\name{summarize.model.parameter}
\alias{summarize.model.parameter}
\alias{normalize.AIC}
\title{Summarize fitted model}
\usage{
summarize.model.parameter(
  params,
  time = 2000,
  scale = 1,
  jitter = 0,
  min.value = 0
)

normalize.AIC(AIC, scale = 1, sample.size = 6667)
}
\description{
Summarize fitted model for properly scaled AIC and parameters
}
