% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_Gaussian_decomposition.R
\name{mix_errors_by_weight}
\alias{mix_errors_by_weight}
\title{Mix error distributions}
\usage{
mix_errors_by_weight(errors, weights)
}
\arguments{
\item{errors}{a list of error distributions (a list of data frames)}

\item{weights}{a numerical vector of probability weights}
}
\value{
The merged error distribution (a data frame)
}
\description{
Mix error distributions by their weights
}
