% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade_hidden_state_prediction.R
\name{predTraitByPIC}
\alias{predTraitByPIC}
\title{Predict hidden state under the BM model}
\usage{
predTraitByPIC(phy, trait, rate = 1)
}
\arguments{
\item{phy}{phylo-class object from \code{ape} package}

\item{trait}{named vector of tip trait values; names should match the tip labels; missing values are hidden states}

\item{rate}{the rate of BM}
}
\value{
a data frame listing the means and variances of the hidden states
}
\description{
Predict hidden state under the BM model given a phylogeny, some known states, and the rate of BM
}
\details{
this function is modified from the function in \code{picante} package
}
